+++ gamesman/src/mdnb.cc
@@ -35,7 +35,7 @@ EXTERNC {
 
 POSITION gNumberOfPositions  =  0;
 POSITION gInitialPosition    =  0;
-POSITION kBadPosition        = -1; /* This can never be the rep. of a position */
+POSITION kBadPosition        = (unsigned long long) -1; /* This can never be the rep. of a position */
 
 STRING   kAuthorName         = "Brian Carnes";
 STRING   kGameName           = "Dots and Boxes";
